+++ src/main/java/chess/domain/pieces/EmptyPiece.java
@@ -1,7 +1,7 @@
 package chess.domain.pieces;
 
-import chess.domain.Team;
-import chess.domain.math.Direction;
+import chess.domain.pieces.component.Team;
+import chess.domain.Direction;
 import chess.domain.pieces.component.Name;
 
 public final class EmptyPiece extends Piece {
