+++ src/main/java/chess/domain/pieces/Knight.java
@@ -1,12 +1,12 @@
 package chess.domain.pieces;
 
-import chess.domain.Team;
-import chess.domain.math.Direction;
+import chess.domain.pieces.component.Team;
+import chess.domain.Direction;
 import chess.domain.pieces.component.Name;
 
 import java.util.List;
 
-import static chess.domain.math.Direction.KNIGHT;
+import static chess.domain.Direction.KNIGHT;
 
 public class Knight extends Piece {
