+++ src/main/java/chess/domain/pieces/Position.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.pieces;
 
 import chess.domain.board.File;
 import chess.domain.board.Rank;
