+++ src/main/java/chess/domain/pieces/component/Team.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.pieces.component;
 
 public enum Team {
     BLACK,
