+++ src/main/java/chess/domain/state/White.java
@@ -1,7 +1,7 @@
 package chess.domain.state;
 
-import chess.domain.Position;
-import chess.domain.Team;
+import chess.domain.pieces.Position;
+import chess.domain.pieces.component.Team;
 import chess.domain.board.Board;
 import chess.view.Command;
