+++ src/test/java/chess/domain/pieces/KingTest.java
@@ -1,7 +1,6 @@
 package chess.domain.pieces;
 
-import chess.domain.Position;
-import chess.domain.Team;
+import chess.domain.pieces.component.Team;
 import chess.domain.board.File;
 import chess.domain.board.Rank;
 import org.assertj.core.api.Assertions;
